requirements.txt
SQLAlchemy==1.3.3
python-dotenv==0.10.2
coloredlogs==10.0
humanfriendly==4.18
mysql-connector-python==8.0.16
mysqlclient==1.4.2.post1
numpy==1.16.3
pkg-resources==0.0.0
protobuf==3.7.1
scipy==1.2.1
six==1.12.0

